+++ bin/florestad/Cargo.toml
@@ -25,6 +25,9 @@ console-subscriber = { version = "0.4.1", optional = true }
 floresta-node = { path = "../../crates/floresta-node", default-features = false }
 bitcoin = "0.32"
 tracing = "0.1.41"
+tracing-appender = "0.2.3"
+tracing-subscriber = { version = "0.3.20", features = ["chrono", "ansi", "env-filter"] }
+dirs = "4.0.0"
 
 [target.'cfg(unix)'.dependencies]
 daemonize = { version = "0.5.0" }
